Revised #168 and made the necessary bug fixes/changes
Fixed bugs appearing after this PR (Delete simple/smart was not working,
spinner was not working other than layout, disconnected nodes was being
removed most probably this was not a bug but it is not something we
expect in SBGN viewer).

Keep a similar file structure to ChiSE.js

Solved 'https://github.com/iVis-at-Bilkent/chise.js/issues/82' in viewer
as well.