Removed 'action-functions' folder
I realized that we do not need seperate files for each
'action-functions' in ChiSE as well (As these files are not so big they
can be merged). Now there is no 'action-functions' folder instead of it
I put all 'action-functions' into 'undo-redo-action-functions.js' file.
(In viewer there is only 'add remove action functions' section in this
file yet)